admin_panel/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
streamlit==1.39.0
firebase-admin==6.5.0
google-cloud-firestore==2.17.0
python-dotenv==1.0.1
requests==2.32.3
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.4.0
email-validator==2.2.0
pydantic[email]==2.9.2
pandas==2.2.3
plotly==5.24.1
matplotlib==3.9.2
streamlit-aggrid==1.0.5
ffmpeg-python==0.2.0
aiofiles==24.1.0
pytest==8.3.3
watchdog==5.0.2